Fix incorrect token processing
In the lexer, the returned tokens have the incorrect source chunk, since the
offset is the position in the source file at the start of the token.  The
result is that the first token have an empty chunk, and all the other chunks
are associated to the wrong token.

Implement a second pipeline stage to add the correct source chunk for each
token, using the offset of the previous token.

Improve documentation.