Update lexer to return more useful informations
The Token type emitted by the lexer contained the chunk field, with source code
around a Go token. However this data is not convenient to use.

Update the Token type by replacing the chunk field with the code and whitespace
fields.  The code contains well defined, consistent and documented data.

Improve documentation.