Export Token fields and scan function
The lexer.go source file exports the Token type, since the lexer implementation
is generic and can be moved to a separate package. However the Token fields and
the scan function were not exported.

Export the Code, Whitespace and Value fields of the Token type, and the scan
function.  Add missing documentation.