Improve source file generation
When generating a Go source file from tokens returned by the Scan function, the
EOF token is not returned, and the last token does not contain the newline
character.

Update the main function to add the final newline character, so that the
generated file is identical to the source file.