Improve documentation about carriage return
The documentation of the Token type incorrectly stated that Whitespace does not
contain carriage return characters.  However this is incorrect, since
Whitespace string data is obtained directly from source code, and not from the
Go scanner (where whitespace characters are discarded).

Fix the documentation of the Whitespace field of the Token type, by stating
that carriage return characters may be present, if the source code has not been
formatted with gofmt.