Second working draft
Second working draft of the goprint command.
The source code of a single Go file is printed on stdout, HTML formatted.

The HTML is optimized for printing.  In order to keep code simple, header and
footer are implemented using CSS Paged Media Module Level 3.  Unfortunately
this W3C working draft is not supported by browsers, and the princexml tool is
required in order to produce a PDF document.