Improve coding style
In pkg.go, use short declaration inside the if statement, in order to make code
more readable and coherent with idiomatic Go coding style.

In lexer.go, add an empty line after the last var statement at the begin of a
local scope, in order to make code coherent with the author Go coding style.

In main.go, add a comment to the Context type, in order to make golint happy.