Print correctly general comments and raw strings
The formatter for the goprint command did not handled correctly general
comments and raw strings that span multiple lines.  The problem was visible in
the generated HTML code, since a line number was not printed for these
elements.

Implement a second pipeline stage, in the formatter code, to correctly split
general comments and raw strings on multiple lines, if required.