goefmt: store token position in the lexer
The lexer implementation obtains the full position of a token, since it
required both the token offset and the line.  Currently the line is no more
used but the full position may be useful for a future API that may require, as
an example, the token column.

Replace the line and offset fields from the Token type with the full token
position.