packages: allow testing of old cmd/go
Using go get golang.org/dl/go1.x.y, it is possible to test if goprint
supports old versions of Go.  However, invokeGo will always use the
current version of go.

Add the gocmd global variable in invoke.go and tell the invokeGo
function to use it, instead of the hard written "go".
In the test file, set it to the value of the GOCMD environment variable,
if it is not empty.

Here is an example:

  go get golang.org/dl/go1.10.8
  go1.10.8 download
  GOCMD=go1.10.8 go1.10.8 test ./...

Update the README.md file to document that goprint requires at least Go
1.8 (older version are not available for testing).