Show the module date after the module path
Add the Module.Date method in the internal/packages package, to return
the date when the module version was created, or the current date if it
is not available.

Update the CSS to show the module date after the module path, in the
bottom left corner of the page, if it is available.