Add support for modules
Add the -m flag that, when set, will instruct goprint to print all the
packages in the specified module.

Add the Packages field in the Module type, in the internal/packages
package.

Add the LoadModule in the internal/packages package.  The function loads
the specified module and all the packages belonging to it.

Update the README.md file.