Don't render the h elements on print media
Update the CSS for both modules and package to set the display property
to none, for h1, h2 and h3 elements when the media is print.

d501134 (Improve CSS for screen when printing a module) and
b539208 (Improve CSS for screen when printing a package) did not updated
these rules.